Harden PN532 I2C response handling

+++ PN532_I2C/PN532_I2C.h
@@ -32,7 +32,7 @@ private:
         #endif
     }
     
-    inline uint8_t read() {
+    inline int read() {
         #if ARDUINO >= 100
             return _wire->read();
         #else
